`strokeWidth` now correctly accounted for when calculating clock layout.
If a font renders uses strokes to render its content, the stroke width
is centered on the path and so half of it appears on the 'outside' of
the path. We need to take that into account when calculating the
layout of the clock: how it's scaled to fit a given space, and how
each glyph is positioned within that space.

Essentially, increasing stroke width makes the glyph appear larger than
its native size and so it needs to be reduced in scale to compensate and
fit in the same boundary area.

`Options`/`Paints` refactor:
- Moved `Options.strokeWidth` to `Paints.strokeWidth`
  - Also added `Paints.strokeCap` and `Paints.strokeJoin`
- `Paints` now included in `Options.paints`