Fixed error in `Layout` which caused issues with alignment.
Previously, line breaks during measurement were detected via
erroneous geometric assumptions about the `rect` received in the
callback (specifically, that the callback rect `top` position only
changed when there was a new row). That assumption held in most
situations, but not when applying center or bottom vertical alignment to
the Horizontal layout option.

Now uses the same `isLineBreak` method called during `layoutPass`.
This method was previously defined on the `MultilineLayout` subclass,
which has now been merged into the core `Layout` class as there was no
reason for it to exist.