All clock types now support visibility transitions.
This required a significant redesign of `Glyph`, `ClockGlyph`, and their
related abstract base classes.

IO16 handles visibility transitions entirely in its `GlyphRenderer`, as
before and as it was always designed to do. Its transitions can start
and end independently of the current time, and can take as long as they
want to complete.

FORM and IO18 aren't quite so straightforward and each need bespoke
per-key appear/disappear animations. These need to replace the
normal second-to-second animations entirely, and so must be synchronized
to the current time so that animations are not swapped out
mid-transition.

As a result of all this, `BaseClockGlyph` now has two child ABCs which
should be implemented depending on which of the above behaviours is
needed for a particular clock animation:
- `ClockGlyphSynchronizedVisibility`, the behaviour used by FORM
  and IO18: visibility transitions are synchronized with the normal
  flow of the animation. Requested visibility changes are deferred until
  the start of the next second.
- `ClockGlyphDesynchronizedVisibility`, used by IO16: visibility
  transitions are handled separately from the normal animation flow and
  do not need to be synchronized. Visibility changes can begin at any
  moment.