Major refactoring of FORM glyphs.
Introduced `PathImage` and `PathAnimation` (and its implementations
including `KeyframeAnimation`), which enables compositions of
`PathCommand`s to be rendered and interpolated between. As a result,
glyphs can now be defined as a series of keyframes (start and end
states) with interpolation applied automatically.

FormGlyph was previously implemented entirely using raw path operations
and (many) transforms which made editing or adding new states quite
challenging. FormGlyph has now been entirely converted to use the new
`PathAnimation` framework which is a much simpler mental model to deal
with.

Almost all states are handled using simple `KeyframeAnimation`s,
although there remain a couple of animations (`TwoEnter`, for example)
which rely on multiple 'progress' values - preventing simple keyframe
interpolation - and so are defined using a raw `PathAnimation` in much
the same was as it used to be. Such caveats are rare, though, and still
much easier to understand in their isolated and reduced scope.

The addition of `Path` helper methods such as `triangle` and `tetragon`
also improve the expressiveness and legibility of path commands - again,
easier to construct the mental model vs long lists of `lineTo` commands.

Along the way we have been able to fix the longstanding issue of the 0_1
transition stretching outside its intended vertical bounds. Appear and
disappear transitions have also been redesigned with more individuality
for each glyph character.