Moved wallpaper logic to `WallpaperEngineDelegate` for simpler testing.
ClockWallpaperService.ClockEngine is now a wrapper for this class,
passing all events for the delegate to handle.

The delegate class has no dependencies on the android system so can
be used in plain unit tests, allowing us to test wallpaper event handling
without needing to set up an actual wallpaper service.

Also fixed glitchy behaviour when device is rotated - state updates are
now ignored for a few moments after rotation to allow the torrent of
`onVisibilityChange` events passed from the system to calm down.